vcl: fix epollet test for unhandled evts
Argument to vcl_epoll_ctl_add_unhandled_event is often the result of an
and between events and EPOLLET which is larger than u8

Type: fix

+++ src/vcl/vppcom.c
@@ -2830,7 +2830,7 @@ vppcom_epoll_create (void)
 
 static void
 vcl_epoll_ctl_add_unhandled_event (vcl_worker_t *wrk, vcl_session_t *s,
-				   u8 is_epollet, session_evt_type_t evt)
+				   u32 is_epollet, session_evt_type_t evt)
 {
   if (!is_epollet)
     {
